// lib/types.ts
export type Supervisor = {
  id: string;
  name: string;
  phone: string;
  canEditExchangeRate: boolean;
  connectionStatus: 'متصل' | 'غير متصل';
  lastSeen: string;
  password?: string;
  specialization: ('محفظة كاش' | 'انستاباي' | 'وصلني البيت')[];
  status: 'نشط' | 'غير نشط';
};